feature(server) Preload package dependencies. - Moved a few functions relating to the packager from `Common.hs` to `NPM.hs`. - Added preload trigger to `startup`. - Extracted out `packageJSONToDependencies` from `getProjectDependenciesFromPackageJSON`. - Implemented `preloadNPMDependencies` to lookup dependencies and cache them. - Added `logException` utility function. - Included `async` project in server.
