+++ server/package.yaml
@@ -27,6 +27,7 @@ flags:
 dependencies:
   - aeson ==2.0.3.0
   - aeson-pretty
+  - async
   - aws
   - base >= 4.7 && < 5
   - base16-bytestring
