+++ server/src/Utopia/Web/Logging.hs
@@ -5,8 +5,12 @@
 
 module Utopia.Web.Logging where
 
+import           Data.Text
 import           Protolude
 import           System.Log.FastLogger
 
 loggerLn :: FastLogger -> LogStr -> IO ()
 loggerLn logger mainStr = logger (mainStr <> "\n")
+
+logException :: FastLogger -> SomeException -> IO ()
+logException logger e = loggerLn logger (toLogStr ("Exception: " <> show e :: Text))
