+++ server/src/Utopia/Web/Server.hs
@@ -24,6 +24,7 @@ import           System.Log.FastLogger
 import           System.TimeManager
 import           Utopia.Web.Executors.Common
 import           Utopia.Web.Logging
+import           Utopia.Web.Packager.NPM
 import           Utopia.Web.ServantMonitoring
 import           Utopia.Web.Types
 import           Utopia.Web.Utils.Files
